MAINT: rename _files as files
