+++ pysat/__init__.py
@@ -102,7 +102,7 @@ from pysat import utils
 from pysat._constellation import Constellation
 from pysat._instrument import Instrument
 from pysat._meta import Meta
-from pysat._files import Files
+from pysat.files import Files
 from pysat._custom import Custom
 from pysat._orbits import Orbits
 from pysat import instruments
