+++ pysat/instruments/methods/general.py
@@ -48,7 +48,7 @@ def list_files(tag=None, sat_id=None, data_path=None, format_str=None,
 
     Returns
     --------
-    pysat.Files.from_os : (pysat._files.Files)
+    pysat.Files.from_os : (pysat.files.Files)
         A class containing the verified available files
 
     Examples
