+++ pysat/instruments/methods/nasa_cdaweb.py
@@ -312,7 +312,7 @@ def list_remote_files(tag, sat_id,
 
     Returns
     --------
-    pysat.Files.from_os : (pysat._files.Files)
+    pysat.Files.from_os : (pysat.files.Files)
         A class containing the verified available files
 
     Examples
@@ -353,13 +353,13 @@ def list_remote_files(tag, sat_id,
 
     # Parse the path to find the number of levels to search
     format_dir = dir_split[0]
-    search_dir = pysat._files.construct_searchstring_from_format(format_dir)
+    search_dir = pysat.files.construct_searchstring_from_format(format_dir)
     n_layers = len(search_dir['keys'])
 
     # only keep file portion of format
     format_str = dir_split[-1]
     # Generate list of targets to identify files
-    search_dict = pysat._files.construct_searchstring_from_format(format_str)
+    search_dict = pysat.files.construct_searchstring_from_format(format_str)
     targets = [x.strip('?') for x in search_dict['string_blocks'] if len(x) > 0]
 
     remote_dirs = []
@@ -415,14 +415,14 @@ def list_remote_files(tag, sat_id,
 
     # parse remote filenames to get date information
     if delimiter is None:
-        stored = pysat._files.parse_fixed_width_filenames(full_files,
+        stored = pysat.files.parse_fixed_width_filenames(full_files,
                                                           format_str)
     else:
-        stored = pysat._files.parse_delimited_filenames(full_files,
+        stored = pysat.files.parse_delimited_filenames(full_files,
                                                         format_str, delimiter)
 
     # process the parsed filenames and return a properly formatted Series
-    stored_list = pysat._files.process_parsed_filenames(stored,
+    stored_list = pysat.files.process_parsed_filenames(stored,
                                                         two_digit_year_break)
     # Downselect to user-specified dates, if needed
     if start is not None:
