+++ pysat/instruments/superdarn_grdex.py
@@ -169,7 +169,7 @@ def list_files(tag='north', sat_id=None, data_path=None, format_str=None):
 
     Returns
     --------
-    pysat.Files.from_os : (pysat._files.Files)
+    pysat.Files.from_os : (pysat.files.Files)
         A class containing the verified available files
     """
