+++ pysat/instruments/sw_dst.py
@@ -144,7 +144,7 @@ def list_files(tag=None, sat_id=None, data_path=None, format_str=None):
 
     Returns
     --------
-    pysat.Files.from_os : (pysat._files.Files)
+    pysat.Files.from_os : (pysat.files.Files)
         A class containing the verified available files
 
     Notes
