+++ pysat/instruments/sw_f107.py
@@ -193,7 +193,7 @@ def list_files(tag=None, sat_id=None, data_path=None, format_str=None):
 
     Returns
     --------
-    pysat.Files.from_os : (pysat._files.Files)
+    pysat.Files.from_os : (pysat.files.Files)
         A class containing the verified available files
 
     Notes
